+++ db/migrate/20170724080021_events.rb
@@ -4,7 +4,7 @@ class Events < ActiveRecord::Migration[5.1]
   		t.column(:host_id, :integer)
   		t.column(:name, :string)
   		t.column(:description, :string)
-  		t.column(:number_of_tickets, :integer)
+  		t.column(:number_of_tickets, :bigint)
 
   		t.timestamps()
   	end
